requirements.txt
fastapi
uvicorn[standard]
requests
google-auth
python-slugify
python-dotenv
google-adk
litellm
google-cloud-aiplatform>=1.66.0
google-generativeai
Pillow>=10.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
grpcio>=1.59.0
